+++ certidude/api/attrib.py
@@ -4,8 +4,9 @@ import re
 from xattr import setxattr, listxattr, removexattr
 from certidude import push
 from certidude.decorators import serialize, csrf_protection
-from certidude.firewall import whitelist_subject
 from certidude.auth import login_required, authorize_admin
+
+from .utils.firewall import whitelist_subject
 
 logger = logging.getLogger(__name__)
