+++ certidude/api/request.py
@@ -10,12 +10,12 @@ from base64 import b64decode
 from certidude import config, push, errors
 from certidude.auth import login_required, login_optional, authorize_admin
 from certidude.decorators import csrf_protection, MyEncoder
-from certidude.firewall import whitelist_subnets, whitelist_content_types
 from datetime import datetime
 from oscrypto import asymmetric
 from oscrypto.errors import SignatureError
 from xattr import getxattr
 from .utils import AuthorityHandler
+from .utils.firewall import whitelist_subnets, whitelist_content_types
 
 logger = logging.getLogger(__name__)
